Current Working changes
Other Than the Boundry_box.py utility and the elevator zero code, everything else has been tested. The Boundry box needs testing, and has not incorporated the claw just yet. the Zero code does not fully zero the elevator, and only sets the current position of the elevator as its zero, meaning if the elevator is not at the bottom of the elevator, it could potentially extend past its desired length.